+++ PSJira/Functions/Set-JiraIssue.Tests.ps1
@@ -43,6 +43,7 @@ InModuleScope PSJira {
             defParam 'Summary'
             defParam 'Description'
             defParam 'Assignee'
+            defParam 'Label'
             defParam 'Fields'
             defParam 'Credential'
             defParam 'PassThru'
@@ -68,6 +69,8 @@ InModuleScope PSJira {
                     'Name' = 'username'
                 }
             }
+
+            Mock Set-JiraIssueLabel {}
 
             It "Modifies the summary of an issue if the -Summary parameter is passed" {
                 { Set-JiraIssue -Issue TEST-001 -Summary 'New summary' } | Should Not Throw
@@ -98,6 +101,11 @@ InModuleScope PSJira {
                 Assert-MockCalled -CommandName Invoke-JiraMethod -ModuleName PSJira -Times 1 -Scope It -ParameterFilter { $Method -eq 'Put' -and $URI -like '*/rest/api/2/issue/12345/assignee' -and $Body -like '*name*username*' }
             }
 
+            It "Uses Set-JiraIssueLabel with the -Set parameter when the -Label parameter is used" {
+                { Set-JiraIssue -Issue TEST-001 -Label 'test' } | Should Not Throw
+                Assert-MockCalled -CommandName Set-JiraIssueLabel -ModuleName PSJira -Times 1 -Scope It -ParameterFilter { $Set -ne $null }
+            }
+
             It "Updates custom fields if provided to the -Fields parameter" {
                 Mock Get-JiraField {
                     [PSCustomObject] @{
@@ -105,10 +113,12 @@ InModuleScope PSJira {
                         'ID'   = $Field;
                     }
                 }
-                { Set-JiraIssue -Issue TEST-001 -Fields @{'customfield_12345'='foo'; 'customfield_67890'='bar'} } | Should Not Throw
+                { Set-JiraIssue -Issue TEST-001 -Fields @{'customfield_12345'='foo'; 'customfield_67890'='bar'; 'customfield_111222'=@(@{value='foobar'})} } | Should Not Throw
                 Assert-MockCalled -CommandName Invoke-JiraMethod -ModuleName PSJira -Times 1 -Scope It -ParameterFilter { $Method -eq 'Put' -and $URI -like '*/rest/api/2/issue/12345' -and $Body -like '*customfield_12345*set*foo*' }
                 Assert-MockCalled -CommandName Invoke-JiraMethod -ModuleName PSJira -Times 1 -Scope It -ParameterFilter { $Method -eq 'Put' -and $URI -like '*/rest/api/2/issue/12345' -and $Body -like '*customfield_67890*set*bar*' }
-            }
+                Assert-MockCalled -CommandName Invoke-JiraMethod -ModuleName PSJira -Times 1 -Scope It -ParameterFilter { $Method -eq 'Put' -and $URI -like '*/rest/api/2/issue/12345' -and $Body -like '*customfield_111222*set*foobar*' }
+        }
+
         }
 
         Context "Input testing" {